requirements.txt
# Core ML
torch>=2.1.0,<2.4.0
torchaudio>=2.1.0,<2.4.0
transformers>=4.40.0,<4.46.0
accelerate>=0.27.0

# ASR
# openai-whisper is pulled in transitively by faster-whisper; no need to pin both
faster-whisper>=1.0.0

# Translation
deep-translator>=1.11.4

# TTS (Hindi) — use the maintained fork; same TTS module, no MSVC build tools needed
coqui-tts>=0.27.0

# Voice Cloning
resemblyzer>=0.1.4
pyannote.audio>=3.1.1,<3.4.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
noisereduce>=3.0.0
pyworld>=0.3.4
praat-parselmouth>=0.4.3
ffmpeg-python>=0.2.0

# Evaluation
pesq>=0.0.4
pystoi>=0.3.3
jiwer>=3.0.3

# Utils
tqdm>=4.66.0
pyyaml>=6.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
jupyter>=1.0.0
python-dotenv>=1.0.0

# Manual installs (see TASKS.md):
# Wav2Lip: https://github.com/Rudrabha/Wav2Lip
